+++ Changes.md
@@ -19,6 +19,7 @@ Improvements
 - Cancellation : Improved responsiveness of cancellation of tasks waiting on results from another thread.
 - ImageReader : Channels from EXR subimages named `rgb`, `rgba` or `depth` (either uppercase or lowercase) are now loaded into Gaffer's primary image layer.
 - PathFilter : Removed `scene:filter:inputScene` and `scene:path` variables from the context used to evaluate the `paths` plug. This improves performance in some scenarios, and prevents the creation of invalid filters where the paths themselves depend on the current scene location.
+- NodeEditor : Added option in node editor context menu to set interpolation of an animated value.
 
 Fixes
 -----